Refactor out interaction logic into separate modules
